requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==2.2.3
docopt==0.6.2
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
importlib-metadata==0.18
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
locustio==0.11.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.1
packaging==19.0
pkg-resources==0.0.0
pluggy==0.12.0
py==1.8.0
pylint==2.3.1
pyparsing==2.4.0
pytest==5.0.1
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
zipp==0.5.2
